pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "knowai"
version = "0.3.0"
license = {file = "LICENSE"}
authors = [
  { name="Chris R. Vernon", email="chris.vernon@pnnl.gov" },
]
description = "A conversational RAG agent pipeline using LangGraph"
readme = "README.md"
requires-python = ">=3.11" # Adjusted based on PyMuPDF and asyncio features
maintainers = [
    {name = "Chris Vernon", email = "chris.vernon@pnnl.gov"},
]
classifiers = [
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
dependencies = [
    "langchain==0.3.26",
    "langchain-core==0.3.67",
    "langchain-community==0.3.26",
    "langchain-openai==0.3.27",
    "langgraph==0.5.0",
    "python-dotenv==1.1.1",
    "PyMuPDF==1.26.1", # fitz
    "faiss-cpu==1.11.0",
    "tiktoken==0.9.0",
    "streamlit==1.46.1",
    "fastapi==0.115.14",
    "pydantic==2.11.7",
    "boto3==1.39.1",
    "uvicorn==0.35.0", 
    
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.12.1",
    "trio>=0.30.0",
]
docs = [
    "setuptools>=57.0.0",
    "build>=0.5.1",
    "ipykernel>=6.15.1",
    "autodoc>=0.5.0",
    "nbsphinx~=0.8.6",
    "sphinx~=7.2.6",
]
deploy = [
    "twine>=4.0.1",
]

[project.urls]
Repository = "https://github.com/crvernon/knowai"

[tool.hatch.build]
packages = ["knowai"]  # set importable package name to "cerf"
include = ["knowai"]
